New functions for performing HTTP authentication.
These functions can be used by a view in order to allow clients to
authenticate themselves with a username and password (rather than
using the normal Django user authentication system with a session ID
cookie.)

In particular, we want to provide a simple, copy-and-pasteable "wget"
command that users can run to download one or more files, and that
command should not *itself* include any passwords or authentication
tokens (rather, the user should be prompted for their password
interactively when they run the command.)

For various reasons, we don't want to allow web browsers to use HTTP
authentication, even if the user stumbles upon a URL that allows it:

- most browsers don't present HTTP authentication prompts in a very
  user-friendly way

- most browsers keep HTTP credentials cached, without showing any
  indication of the fact, or providing any way for the user to "log
  out" short of closing the browser

- random clients with JavaScript enabled could be hijacked into a
  distributed brute-force password guessing attack

- older implementations are ambiguous in their handling of non-ASCII
  passwords, which could lead to confusing authentication failures

In particular, the risk of browser hijacking is the reason the normal
Django login process is so convoluted (requiring multiple cookies,
hidden form fields, Referer headers...) which makes authentication
painful for programmatic access.

Thus, the http_auth_allowed function permits authentication only for
whitelisted user agents (currently wget and libwfdb.)  It's fine to
add additional UAs to this list, but before doing so we should verify
that:

(a) the UA in question handles passwords (and password-protected data)
    in a relatively sane and secure manner

(b) the UA doesn't provide a mechanism for untrusted scripts to issue
    HTTP requests on the user's behalf

For a view to enable HTTP authentication requires two steps:

- The view must check the provided credentials (to authenticate the
  user) by calling check_http_auth().

- If the user does not have permission to access the given resource,
  the view must call require_http_auth() to return a "401
  Unauthorized" response.

(Note that many documents, including the HTTP standards, use the words
"authentication" and "authorization" confusingly and often
interchangeably.  "401 Unauthorized" really means that the user is
not *authenticated* - they haven't provided credentials, or the
credentials they provided were incorrect.  If the user is known but
doesn't have permission to access a resource, they should receive a
"403 Forbidden" response.)