display_project_file: handle NotADirectoryError.
When visiting a URL such as:

http://localhost:8000/content/demobsn/1.0/RECORDS/foo

where 'RECORDS' is a regular file and not a directory, this should
result in a 404 Not Found response.  However, in this case, open()
raises a NotADirectoryError (EISDIR) rather than FileNotFoundError
(ENOENT).

Thus, we want to catch both of these exception types and raise Http404
in both cases.