Adds option for editors/authors to download all files
The functionality to people download all the files for protected projects
exists, the only issue is that it works for published projects.

Here I add the functionality for active projects also. If the person is
either the editor or one of the authors (which they should not have the
link, but still is possible for them because is their project), then
they can download all the files from the project, ether to have them or
to do the "editing".

Here I use a lot of the groundwork made by Benjamin, and even it might
be possible to eventually merge with "serve_published_project_file"
since it uses basically the same things.