Removes approved_users duplicates DUASignature
There is a duplication on how the projects with Licenses  and Credentialing
handles access. The DUASignature has a bit more information than
approved_users, it has the date of when it was signed.

Here I removed the duplication by removing approved_users from the project
since its there in the DUASignature.