Adding an abort on the user search
The user search is a bit buggy because every time you type a letter
it does a request. Since the request are asynchronous, they can be swapped
or just display all the users.

This is an attempt to abort the previous request, and start another.
Also, I am adding an icon that says if the account is active.

This way once a person complains about the registration email, we can
easily search here.