Mark the anonymousaccess cookie as Secure, HttpOnly, and SameSite.
Secure means the cookie must not be transmitted over unencrypted
channels.

HttpOnly means the cookie must not be visible to JavaScript.

SameSite means the cookie must not be transmitted for "cross-origin"
requests, such as when an image, script, or stylesheet is referred to
by another website.  ("Lax" means that following a link is not treated
as a cross-origin request.)  This is fairly new and not universally
supported, but dramatically improves security when it is supported.