After HTTP authentication, save login state using session cookies.
Verifying passwords is (*by design*) a slow operation, and if someone
is downloading a large number of files using wget, the time spent
verifying their password over and over can be a significant
performance bottleneck.

Instead of doing this, set a cookie that the client can use to
authenticate itself for future requests.  This is done using the
existing Django session mechanism to securely store the current user
ID and a hash of the hashed password, just as the regular login page
does.

(Once a variable has been defined in request.session, it causes the
session middleware to assign a persistent session ID, and return it as
a cookie with the response.  Thus, this doesn't require any changes to
views that invoke check_http_auth.)

To minimize the possible security impact of this change, this
essentially replicates the behavior of auth.login(), without actually
using auth.login(), so that a client that authenticates in this way is
not then permitted to access other URLs that require "login"
authentication.