Only create sessions for clients that support cookies.
Sessions are currently stored on the server side, so if a client
doesn't support cookies, we would like to avoid creating a new session
for every request.  (Requests will of course be slower for such
clients since the password must be verified every time.)

To implement this, set a test cookie when delivering the initial 401
Unauthorized response.  (We don't use the "sessionid" cookie itself
for this purpose, because the session middleware may try to delete
that cookie automatically if it's invalid.)

Only if the request includes at least one cookie, as well as a valid
username/password, do we save the authentication data (which triggers
a session to be saved on the server side.)