Python requirementes updates
There is a "low" security issue on Pillow for images.

"An issue was discovered in Pillow before 6.2.0. When reading specially crafted invalid image files, the library can either allocate very large amounts of memory or take an extremely long period of time to process the image."

Here I do an update to that and Django.
Also, for some reason my psycopg2-binary hash was incorrect, so I add one for OSX.