Add InlineFileView class.
This file view displays the document as an inline frame.  It's
intended specifically for PDF documents and probably isn't well suited
for much of anything else.

Local CSS rules are used to set the height of the iframe so that (in
most cases) when the page is first loaded, the entire iframe should be
visible plus a small white border at the bottom; the page footer
should be off the screen.  These rules could perhaps be improved with
some complicated flex stuff, but I'm not sure it's worth it.