physionet_nginx.conf: set Content-Security-Policy header.
This header should preferably be set for all project files, but
unfortunately needs to be turned off for PDF files in order to
accomodate Chromium users.

Note that if this header is returned by uwsgi in an X-Accel-Redirect
response, nginx doesn't pass it on to the client by default.

For public projects (either accessed via /files or via
/static/published-projects), set the header for all non-PDF files.

For protected/active projects, preserve the header (or lack thereof)
that is sent by uwsgi.