Simplify the view for the past credentialing applications
The function for past credentialing applications was getting a bit complicated.
Here I create another function that manages part of the function.

I move out the custom search for a specific application, and the default
search for all the applications on to a separate function and return
two arrays that would contain the final result for the successful applications
and the unsuccessful applications.