Trying to fix server 502 error
There was an error in which uWSGI stopped working with no errors in the log and
just stopped responding. To make it work again it was restarted, but it happened
several times after that.

This is an attempt to fix that error, adding a bit more logging, and enabling
"thunder-lock" to see if that was the cause of the problem or part of it.
Enabling thunder-lock should put a stop to the thundering herd problem if we have it.

Also, here I introduce the concept of worker recycling.
Worker recycling can prevent issues that become apparent over time such as memory
leaks or unintentional states. In some circumstances, however, it can improve
performance because newer processes have “fresh memory space”.

uWSGI provides multiple methods for recycling workers. Assuming the app is relatively
quick to reload, all three of the methods below should be effectively harmless and
provide protection against different failure scenarios.