Replacing the static content of the Challenge page
Here I replace the static content of the static page, and also add the
published papers and contributed software section.

For this to work, there has to be some manual labor done.

For the software section, the folder 'sources' has to exists under the project.
If the folder exists, it will create a URL under the challenge index page to a
Nginx http index page, just like in the old website.

For the papers, is somewhat similar.
- Create a papers folder and put all of the papers into the folder (e.g. like the papers folder for https://physionet.org/content/challenge-2018/)
- Save https://archive.physionet.org/challenge/2018/papers/ as an HTML file and upload the file to the challenge files named 'index.html' and place it in the same folder.