New functions edit_log_history, copyedit_log_history.
These functions obtain a list of EditLogs or CopyeditLogs for a
project, respectively, in chronological order.

In contrast, the edit_logs and copyedit_logs properties yield
unordered collections of objects.

edit_log_history and copyedit_log_history should be used in place of
edit_logs and copyedit_logs when the order matters (i.e. when
displaying a timeline of events.)