Use edit_log_history and copyedit_log_history where appropriate.
In several places, the "submission history" of a project is displayed:

* /projects/*/submission/

  For active projects; displayed to the author.  Corresponds to
  project.views.project_submission.

* /console/submitted-projects/*/edit/#timeline

  For active projects; displayed to the editor.  Corresponds to
  console.views.edit_submission, which invokes Metadata.info_card.

* /projects/rejected/*/submission-history/

  For archived projects; displayed to the author or admin.
  Corresponds to project.views.rejected_submission_history.

* /projects/published/*/*/submission-history/

  For published projects; displayed to the author or admin.
  Corresponds to project.views.published_submission_history.

All four cases were previously invoking 'project.edit_logs.all()' and
'project.copyedit_logs.all()' to obtain records of the submission
history.  Since this doesn't guarantee any particular order, the
results (in cases where a project was submitted multiple times) could
be scrambled in various ways.

(Confusingly, the displayed timeline was not *completely* scrambled,
because the events in the timeline don't map one-to-one to EditLog or
CopyeditLog objects.)

In all four of the above cases, use 'edit_log_history' and
'copyedit_log_history' to ensure that the events are shown in the
desired order (but note that this still assumes that all CopyeditLogs
occur after all EditLogs.)