Add 'getmigrationtargets' management command.
This command inspects the available migrations in the source
directory, as well as the list of migrations that have already been
applied, and reports a set of "targets" that can be passed to
'manage.py migrate' or 'manage.py migratetargets'.

Conceptually, migrations are divided into:

- migrations that have already been applied ("current")

- migrations that are not yet applied and are prerequisites for
  upgrading the server ("early")

- migrations that are not yet applied and must not be applied until
  after the server is upgraded ("late")

Distinguishing "early" from "late" migrations must be done by
labelling the Migration class with a custom attribute
(MIGRATE_AFTER_INSTALL).

This command determines what needs to be done in order to:

(a) apply all "current" migrations, but no "early" or "late"
    migrations (manage.py getmigrationtargets --current)

(b) apply all "current" and "early" migrations but no "late"
    migrations (manage.py getmigrationtargets --early)

(c) apply all "current", "early", and "late" migrations
    (manage.py getmigrationtargets)

In each case, it displays the arguments that would need to be passed
to the 'migrate' command in order to achieve the desired state.  These
can then be applied by hand, or applied as a batch by the
'migratetargets' command.

The output of this command can also be useful in enabling scripts to
check whether or not any migrations need to be applied.