Allow the 'pn' user to reload/restart servers.
The pn user is allowed to trigger a reload of the uwsgi daemon, using:

    sudo systemctl reload emperor.uwsgi.service

This will stop all existing uwsgi worker processes and launch new ones
to handle new requests.  The systemctl command will wait until all of
the old processes have exited (although this can be fooled if the
command is issued multiple times.)

This command does not start the uwsgi daemon if it isn't currently
running.

The pn user is also allowed to trigger a restart of the
django-background-tasks daemon, using:

    sudo systemctl try-restart django-background-tasks.service

This will immediately interrupt the process (interrupting any running
tasks) and then restart it.  This command does not start the daemon if
it isn't currently running.