New git hooks to automatically install new versions when pushed.
When a new version is to be installed, numerous things need to happen,
in order:

(a) new dependencies to be downloaded and installed
(b) early migrations perfomed on the database
(c) new static files installed into /data/pn-static
(d) new server code unpacked into /physionet/physionet-build
(e) uwsgi server, and/or django-background-tasks, reloaded
(f) late migrations performed on the database

To allow automated testing of the upgrade process, and to reduce the
possibilities for human error, we'd like all of this to be done
automatically.

Furthermore, all of the results of these commands should be logged in
case of problems (the status messages sent back to the pusher should
be kept to a minimum).  And, if we can detect ahead of time that any
of these commands will fail, the push should be rejected.

In the past, steps (a), (c), and (d) were done by the post-receive
hook; steps (b), (e), and (f) were ignored and needed to be done
manually.  Here, we add the remaining steps:

- Step (a) is now performed by the update hook, so that if it fails,
  the push is rejected.

- The lists of target migrations are generated (by 'manage.py
  getmigrationtargets') during the update hook, so that if there are
  any inconsistencies, the push is rejected.

- Step (b) is now performed by the post-receive hook (using 'manage.py
  migratetargets').

  This is not done in the update hook because, if it *partially*
  succeeds, we do not want to be left with migrations that have been
  applied but are not in the git repository.

- Step (e) is now performed by the post-receive hook (using
  'systemctl' commands which are authorized via
  /etc/sudoers.d/physionet).

  django-background-tasks is only restarted if there are any
  migrations being applied.  In most other cases, it should not be
  necessary, and we want to avoid interrupting long-running tasks if
  possible.

- Step (f) is now performed by the post-receive hook (using 'manage.py
  migrate').