Added Pinned news to console and home
I am adding a removable alert to the PhysioNet home page, and also
a way to manage this via the console.

Only 1 pinned news is allowed at the time, and to add a pin News
one must either edit an existing and set pinned true, or create
a new News object, and set pinned to true.