Added variables for DOI assignment
Since the DOI will be assigned by the API, we need the authentication variables to be added to the environment file.

To use the API, the two links are:
 - https://api.datacite.org/dois
 - https://api.test.datacite.org/dois

To edit the an already API, the URL changes slightly, but that will be introduced in a future commit in the code.

To see the information in the web browser, the two links are:
 - https://doi.datacite.org/
 - https://doi.test.datacite.org/

At the moment the credentials for both are the same. This is something that we might want to change in a future time.