Added functions to generate and publish DOI
There are two new functions in the console utility.
The first one to generate a DOI draft with the required information.
The second to publish that DOI and update it with the findable URL.

To updates DOI, the URL for the API changes slightly, you can see the modified URL in the code.

This is commit only uses basic information to generate the DOI.
 - Authors
 - Title
 - Publisher
 - Publication year
 - General Resource Type
 - URL
 - Description
 - Abstract

More information can be added in the future.

The test ommits the DOI publishment but it creates a draft DOI.