set past applications order by decision date not application date
If the last rejected application was created a month ago, it wont
be shown in the top of the past applications, but in the time that
the application was created. If one wanted to make changes to that
one would have to try and remember what was the date of something
about that application.

Here I set the proper order for past applications.