Fix sense of condition for restarting django-background-tasks.
In the post-receive hook, we check to see if there are any new
migrations; if not, we don't restart django-background-tasks
automatically.  This check was broken in commit
c3460e556fbc912f7404fb1e441a8c8ecb08436a ("Add support for push
options in the git hooks").

If "cmp -s OLD-TARGETS LATE-TARGETS" succeeds, then the two files are
identical, thus there are no migrations to be applied, thus there is
no need to restart background tasks.  If the "cmp" command fails, then
we do want to restart background tasks.