Simplified variables and resource types
I am removing the hardcoded resource types and setting by default
that the type of DOI will be a Dataset. Then it checks if the
resource type is software and do the changes if needed.

Also I am simplifying the variables setting by default the TEST
DataCite variables, and if the production configuration is used,
then it gets the production variables from the env file.