Add message if DataCite is not configured
A simple message above the DOI read-only box will be displayed
if the DataCite prefix is not in the .env file.

This is useful for people reusing the platform, and for us to see
why there is NO DOI.

I am assuming that if you took the time to set the prefix, you
also took the time to set the user/password.