Create draft on ActiveProject "Accepted"
When the editor marks the project as accepted, before the copy edit
part, if the editor marks auto_create DOI then it will be generated.

If the CoreProject doesn't already have a DOI and the DOI was
created, then a second DOI is created and assigned to it.

If I recall correctly, we agreed that each project will have two DOIs from the beginning.
  - One for the each version
  - One for the latest version

IF the editor decides on not automatically creating the DOI, this
will have to be done manually later.