project_content: disable fields if not editable.
In /projects/*/content/, if the user is not the submitting author, or
if the project is currently under review, the user is not permitted to
change any of the input fields.

Furthermore, if the project is a new version of an already-published
project, the author is not allowed to change the title.

These things were enforced, to an extent, by loading the script
'disable-input.js' if the appropriate criteria were not met (not
is_submitting or not project.author_editable).  This obviously doesn't
work if scripts are disabled, and it also has no effect on ckeditor
widgets.

The correct way to mark fields as disabled is to set the HTML
"disabled" attribute.  To do this, we add an optional 'editable'
parameter to the ContentForm constructor; if editable is False, we set
the Django disabled attribute on each form field.

(The constructor's catchall *args parameter is removed, because no
callers currently use positional arguments, and they shouldn't -
BaseModelForm currently has 11 optional positional arguments and if
you can remember what order they come in, your memory is much better
than mine.)

(Note that in addition to setting the HTML disabled attribute, marking
a Django form field as disabled means the submitted value, if any,
will be ignored when saving the form.  This is a useful safeguard,
although it should not have any impact at present - we already forbid
POST requests in those cases.)

We leave the 'disable-input.js' script in place for the time being, as
a kludge to handle the ReferenceFormSet - there isn't a completely
obvious way to set 'disabled' across a form set.  This should be
fixed.

Finally, if the form is not editable, we hide the submit button to
avoid confusing people.