project_discovery: disable fields if not editable.
In /projects/*/discovery/, if the user is not the submitting author,
or if the project is currently under review, the user is not permitted
to change any of the input fields.

The correct way to mark fields as disabled is to set the HTML
"disabled" attribute.  To do this, we add an optional 'editable'
parameter to the DiscoveryForm constructor; if editable is False, we
set the Django disabled attribute on each form field.

We leave the 'disable-input.js' script in place for the time being, as
a kludge to handle the PublicationFormSet and TopicFormSet - there
isn't a completely obvious way to set 'disabled' across a form set.
This should be fixed.

Finally, if the form is not editable, we hide the submit button to
avoid confusing people.