project_access: disable fields if not editable.
In /projects/*/access/, if the user is not the submitting author, or
if the project is currently under review, the user is not permitted to
change any of the input fields.

The correct way to mark fields as disabled is to set the HTML
"disabled" attribute.  To do this, we add an optional 'editable'
parameter to the AccessMetadataForm constructor; if editable is False,
we set the Django disabled attribute on each form field.

This eliminates the need for 'disable-input.js', and we can also
remove a bit of cruft pertaining to "passphrase_form", which is no
longer part of this page.

Finally, if the form is not editable, we hide the submit button to
avoid confusing people.