Add a basic test case for publishing DOIs.
To automatically generate and publish DOIs, we make requests to the
DataCite web API at 'api.datacite.org'.  In order to test this
automatically, we need to emulate those HTTP requests without actually
accessing the network.

To do this, we create a 'requests_mock.Mocker' object (which
automatically intercepts all requests made using the 'requests'
library, returning the specified HTTP responses to the caller).

As a basic test, we simply need to handle the URLs that will be
accessed in the process of publishing a project, and for each URL,
return a JSON document with the required attributes.