Simplified GCP on console views and added GCP prefix
I created the function bucket_info with an email flag that gets me either the bucket name or the bucket access email
This function will the GCP prefix configured in the settings.

Simplified the bucket creation function and the function to check if the buckets exists.

I use the sites configuration to generate the proper domain for the organization.
This is because we use the GCP organization that uses the format of ****.physionet.org.
That format says that all the buckets under physionet.org belong to us.
Since we dont want to hardcode the words physionet.org I am using the sites ID to get that.

I moved splitted the project management function removing all GCP related actions into a separate function.
This way is easier to read and understand.