Hide file manipulation buttons when files are not editable.
When a project's files (/projects/*/files/) are viewed by an author
other than the submitting author, or the project is currently under
review, the author is not allowed to modify the files or make requests
for additional storage.

To reduce confusion in this situation, it's better to hide all of the
buttons and form fields rather than simply disabling them.  In
addition, the check boxes in the list of files don't do anything, so
these should be hidden as well.

To implement this, we add a new template parameter 'files_editable'
which is true if the author is currently allowed to make changes.

 - In project_files.html, hide the "Request Storage"/"Cancel Request"
   forms if files_editable is false.

 - In project_files_form.html, hide the "Upload Files"/"Create
   Folder"/"Rename Item"/"Move Items"/"Delete Items" forms if
   files_editable is false.

 - In edit_files_panel.html, hide the checkboxes if files_editable is
   false.

This parameter must be set appropriately by each of the views that
directly or indirectly include these three templates:

 - project.views.project_files (only accessible by project authors;
   editable only by the submitting author when the project is
   "author_editable")

 - console.views.copyedit_submission (only accessible by the handling
   editor and only accessible when the project is "copyeditable", so
   the form is always editable)

 - project.views.project_files_panel (accessible both by project
   authors and by editors, editable depending on the state of the
   project)