project_files.html: clarify warning message.
When /projects/*/files/ is viewed by an author other than the
submitting author, it displays a warning message "Only the submitting
author may edit the metadata".

However, this page is concerned with the project files, not metadata,
so replace the message with "Only the submitting author may upload or
edit files".