Scripts for running test server in a virtual machine.
'install-pn-test-server' is a script to install the demo server, from
git, on a freshly installed Debian system.  Although this system is
meant for testing, it should mimic many aspects of the live server.

'ssh-install-pn-test-server' is a wrapper that invokes
'install-pn-test-server' to install the demo server on a (virtual)
machine via SSH.

'kvm-pntest' and 'kvm-pntest-create' are scripts for creating and
running a VM using qemu/kvm on Debian.