Run doctests in notification.utility and physionet.utility.
The doctest module provides a way to write simple unit tests that also
serve as API documentation (or documentation that also functions as a
test case).

By defining a 'load_tests' function we can integrate these tests with
the Django testing framework, so that 'manage.py test' runs doctests
automatically.

https://stackoverflow.com/a/20564401