Fix MIMIC AWS access grant using cloud formation
This issue is because of AWS and support in China. There was an
initial user that is using AWS in China and when he tried to access
our project using AWS it gave an error.

After we contacted our AWS contact they told us it was because of
the reason above.

There is a way around this error as they suggested, and its to
set another bucket using a Chinese collaborator.