AssociatedEmail: add verification_token field.
In order to verify that someone is able to receive email at a given
address, we must send a secret token that cannot be guessed and is
uniquely associated with that email address.  The easiest way to do so
is to generate a random string and save it in the server-side
database; so here, we add an optional field to the AssociatedEmail
model.