New management command 'sendnewemailtokens'.
When we change the algorithm for verifying email addresses, the old
insecure links that have already been sent to users will no longer
work.

After the changes are deployed on the server, this command will
identify people who have received an insecure link that will no longer
be accepted for verification.  It will send an updated message to each
of those people.