Fix the project version comparison
This can also be used to improve the validation, but its not a validator perse.

A version number consists of two or three dot-separated numeric components,
with an optional "pre-release" tag on the end.  The pre-release tag consists
of the letter 'a' or 'b' followed by a number.

If the numeric components of two version numbers are equal, then one with a
pre-release tag will always be deemed earlier (lesser) than one without.

The following are valid version numbers (shown in the order that would be
obtained by sorting according to the supplied cmp function):

   0.4       0.4.0  (these two are equivalent)
   0.4.1
   0.5a1
   0.5b3
   0.5
   0.9.6
   1.0
   1.0.4a3
   1.0.4b1
   1.0.4

Here I remove the comparison of versions, and add a function that sorts
the project in version order.

Then it sets the latest version and sets the version order of the published
projects.

This only happens  when a project is published.