Move temporary storage for uploaded data to /data/www-tmp.
The "client_body_temp_path" is a directory where nginx will store the
body of a request while it is being uploaded.  (We want nginx to
buffer the entire post body before handing it off to uwsgi, because we
don't want to keep a uwsgi thread occupied simply waiting for the data
to be transferred.)

Since we are allowing extremely large posts, especially to
/projects/*/files/, we don't have space in the root (or /var)
partition to store them.  So move these into /data/www-tmp instead.

Note that nginx will automatically create the specified directory if
it doesn't exist.  Note also that nginx *unlinks* each temporary file
as soon as it's created, so the temporary files are not visible in the
filesystem, but they are still taking up space on disk and can be
identified by 'lsof'.