Add an email notification for failed tasks
When a task fails it sends a signal.
There are two signals, rescheduled and failed.

Here I capture both signals and send emails on each one.

Also, the default MAX_ATTEMPTS for this is by default 25.
I am reducing this to 5. In theory, if it fails once it will fail
all 5 times because the code should not change that fast.