Increase selector specificity for .search-input.
The rules for '.search-input' in physionet.css were previously
considered to be a higher priority than the rule for 'input' in
form-control-input.css.  Following commit
2576c6c12cc66576e14d58b468a9c58005eab852, the latter now applies only
to certain types of input elements and thus is considered to be higher
specificity.

Increase the specificity of the rules in physionet.css by changing
'.search-input' to '.navbar .search-input', so that these rules
continue to have higher priority.