Check quota and file size using JS before upload
Here I am adding the size of all the files being uploaded after
the files were chosen but before the form was submitted.

If the files sizes exceeds the remaining quota a message will be
shown. If the form was submitted and the file sizes exceeds the
remaining quota then the message will be shown and the form will
be prevented from running.