Add test-upgrade.sh script.
This script can be used to test the "live upgrade" process from one
revision to another, by running the test suite under varying
conditions.  In particular, since we use mirror databases for testing,
we can run the test suite without applying the default migrations.
Thus, we can run the test suite (for both the old and new revisions)
against a database where "early" migrations have been applied but
"late" migrations have not - simulating a live environment that is in
the process of an upgrade.

For these tests to pass requires migrations to be designed very
carefully.  For example, adding a new non-nullable column to an
existing table typically requires a multi-stage migration:

- an early migration that adds the column, but makes it nullable;

- application code that permits the column to be null, but always sets
  it to a non-null value when saving;

- a late migration that sets a default value for any existing rows and
  alters the column to be non-nullable.

In many cases - especially for migrations that only affect
editors/administrators and not ordinary visitors - this process may
not be strictly necessary in order to achieve a seamless transition.
However, enforcing this process should help us to catch mistakes.