kvm-pntest, kvm-pntest-create: disable KVM if unavailable.
If the host system lacks hardware virtualization (e.g. no CPU support
or disabled by BIOS), or if the kernel module (kvm_amd, kvm_intel)
isn't loaded, or if we don't have permission to use it... we can still
use qemu in emulation mode.  This will be slower than KVM but it
should still work.