request_data_access: move checkbox inside label element.
In /request-access/*/*/, the label for the checkbox is displayed as an
inline block.  Consequently, if the window is too narrow for the
checkbox and label to fit on a single line, the line will be broken
after the checkbox.

Moving the checkbox inside the label makes it appear inline with the
text.