Capture the user registration form integrity error
There is an error where when a person double clicks the user
registration form, it will send two requests to create an account.
Since at the moment, both requests are valid, the uniqueness will pass.

Here I try to catch the integrity error if a person double clicks
the registration button.
It will also not send the second email if the error is triggered.