Reorganize and clarify the Metadata and SubmissionInfo classes.
It is important to distinguish between the meaningful content of a
project (as supplied by the author and possibly modified by the
editor), and the status of the project within the submission system
(which is largely outside the author's control and is not intrinsic to
the resource they are seeking to publish.)

For the most part, the former information has been stored in the
abstract Metadata class, whereas the latter has been stored in the
abstract SubmissionInfo class.  (Both PublishedProject and
UnpublishedProject - and hence ActiveProject and ArchivedProject -
inherit from both of these classes.)

Formalizing the distinction between these two types of data is useful
for tracking the "modification time" of a project, and may also be
useful in the future for tracking "version history".

To be consistent, the following fields which were previously part of
Metadata are now moved into SubmissionInfo:

- creation_datetime
- edit_logs
- copyedit_logs
- version_order
- anonymous

Since moving these fields from one abstract class to another does not
alter the schemas of any *concrete* models, this is not a migration.