UnpublishedProject: conditionally update modified_datetime.
It is useful to know the date and time when a project's content was
last changed.  (For example, this date and time are displayed on the
author's projects page, and are displayed when the project is viewed
via the anonymous access mechanism.)

This timestamp should not be changed when the project's submission
status changes, since that is not (conceptually) a change to the
content of the resource.

The ActiveProject (and ArchivedProject) models, which both inherit
from UnpublishedProject, have a 'modified_datetime' field, which is
automatically set to the current time when it is saved.  For the most
part, this does the right thing - most operations that alter the
project content will invoke save() on the project and thus the
timestamp will be updated.

However, some operations (such as assigning an editor) require saving
the project without changing its content.

To handle all such operations transparently - including operations
performed by scripts, or the Django admin interface - we modify the
behavior of UnpublishedProject.save().

- If the object was not originally loaded from the database,
  modified_datetime is set to the current time.

- If any of the Metadata fields differ from the values that were
  originally loaded from the database, modified_datetime is set to the
  current time.

- If the object was loaded from the database but some or all of the
  Metadata fields were deferred (using QuerySet.defer() or
  QuerySet.only()), modified_datetime is set to the current time and a
  warning is displayed.

- Otherwise, we skip saving all of the Metadata fields as well as the
  modified_datetime field.

(This is done by constructing a list of fields to update based on the
_meta.fields property.  Although this is documented as a private
Django API, _meta.fields appears to be the most accurate way to
determine "the list of fields that save() will save, and that will be
accepted in update_fields".  In contrast, the documented
_meta.get_fields() also returns reverse relations and many-to-many
fields, which are not things that can be "saved" in the same way.)

This behavior can be overridden by explicitly specifying
"update_fields", or by setting "content_modified" as true or false.