Update modified_datetime when associated objects are saved.
The Topic, Publication, and Reference objects represent portions of
the project content that are not saved as part of the
UnpublishedProject object.  As such, the project modification
timestamp should be updated any time these objects are modified.

To ensure that this is done automatically, invoke
project.content_modified() whenever save() or delete() is called on
any of these associated objects.