Copy the modification timestamp when archiving a project.
When a project is archived, it might be useful to keep the
last-modified timestamp for future reference.  Since modified_datetime
is an auto_now field, simply copying the value and then calling save()
will not do this - save() will either set modified_datetime to the
current time (if the content fields are changed) or will leave it
alone (if the content fields are not changed.)

Instead, wait until after copying all of the other fields and
associated objects, and then perform a database update to set the
"correct" modification time on the ArchivedProject.