DiscoveryForm: update project timestamp.
Submitting this form changes visible attributes of a project, which
should be reflected by updating the project modification timestamp.

In particular, this form uses many-to-many fields
(programming_languages, parent_projects), which are not reflected by
saving a model and thus do not cause the project timestamp to be
updated automatically.  Invoke project.content_modified() to ensure
that the timestamp is always updated when this form is saved (even if,
sometimes, it doesn't need to be.)